packages, presentation, checker, loader, cli
the whole project has been split into packages: application, cli, gui,
model.

there is a presentation class that displays the contents of a book to
you in fullscreen mode.

there is a checker class (incomplete) that will check files for bad
formatting.

parts from trainer have been moved to loader, which asks for a file and
loads the book from it.

there is a new cli class that contains the main method, it's mainly
taken from trainer, has however been revised.